refactor: reorganize stop payment types and imports for clarity
