+++ types/checkPayment.ts
@@ -1,7 +1,6 @@
-import { Tags, RelationshipsArray, Relationship, BaseListParams, Sort, BaseCreateRequestAttributes, CheckPaymentCounterparty } from "./common"
+import { Tags, Relationship, BaseListParams, Sort, BaseCreateRequestAttributes, CheckPaymentCounterparty } from "./common"
 import { BasePaymentRelationships } from "./payments"
 
-export type StopPaymentStatus = "Active" | "Disabled"
 export type CheckPaymentStatus = "New" | "Rejected" | "Pending" | "Canceled" | "PendingCancellation" | "InProduction" | "InDelivery" | "Delivered" |
  "ReturnedToSender" | "PendingReview" | "Processed" | "MarkedForReturn" | "Returned"
 
@@ -24,12 +23,7 @@ interface BaseCheckPaymentAttributes {
     amount: number
 
     /**
-     * The status of the stop payment, one of Active or Disabled.
-     */
-    status: StopPaymentStatus
-
-    /**
-     * The checkNumber of the check payments that the stop payment operation will be applied to.
+     * The checkNumber of the check payment.
      */
     checkNumber: string
 
@@ -38,80 +32,6 @@ interface BaseCheckPaymentAttributes {
      */
     tags?: Tags
 }
-
-export interface StopPayment {
-    id: string
-
-    type: "stopPayment"
-
-    attributes: BaseCheckPaymentAttributes
-
-    relationships: {
-        /**
-         * The list of CheckPayments that were stopped by this stopPayment.
-         */
-        stoppedPayments?: RelationshipsArray
-
-    } & Omit<BasePaymentRelationships, "transaction">
-}
-
-export interface AchStopPayment {
-    id: string
-
-    type: "achStopPayment"
-
-    attributes: {
-        /**
-         * The date the resource was created.
-         * RFC3339 format. For more information: https://en.wikipedia.org/wiki/ISO_8601#RFCs
-         */
-        createdAt: string
-
-        /**
-         * Optional. The amount (in cents) above which a payment will be stopped.
-         */
-        minAmount?: number
-
-        /**
-         * Optional. The name of the originator of the payment to look for in payments to stop.
-         */
-        originatorName?: string[]
-
-        /**
-         * The direction of the payments to stop. Debit only.
-         */
-        direction: "Debit"
-
-        /**
-         * Optional. Whether the stop payment can stop more than one payment. False by default.
-         */
-        isMultiUse?: boolean
-
-        /**
-         * Optional. Date only (e.g. "2001-08-15").
-         */
-        expiration?: string
-
-        /**
-         * The description of the stop payment (up to 255 characters).
-         */
-        description: string
-
-        /**
-         * The status of the stop payment, one of Active or Disabled.
-         */
-        status: StopPaymentStatus
-
-        /**
-         * See [Tags](https://developers.unit.co/#tags).
-         */
-        tags?: Tags
-    }
-
-    relationships: Omit<BasePaymentRelationships, "transaction">
-}
-
-export type StopPaymentResource = StopPayment | AchStopPayment
 
 type PendingReviewReasons = "SoftLimit"
 
@@ -237,85 +157,9 @@ export interface CheckPayment {
         memo?: string
 
 
-    } & Omit<BaseCheckPaymentAttributes, "status">
+    } & BaseCheckPaymentAttributes
 
     relationships: BasePaymentRelationships
-}
-
-export interface CreateStopPaymentRequest {
-    type: "stopPayment"
-    attributes: {
-        amount?: number
-        checkNumber: string
-        tags: Tags
-        idempotencyKey?: string
-    }
-    relationships: {
-        account: Relationship
-    }
-}
-
-export interface CreateAchStopPaymentRequest {
-    type: "achStopPayment"
-    attributes: {
-        /**
-         * Optional. The amount (in cents) above which a payment will be stopped. At least one originatorName or minAmount must be provided.
-         */
-        minAmount?: number
-
-        /**
-         * Optional. The name of the originator of the payment to look for in payments to stop. At least one originatorName or minAmount must be provided.
-         */
-        originatorName?: string[]
-
-        /**
-         * The direction of the payments to stop. Debit only.
-         */
-        direction: "Debit"
-
-        /**
-         * Optional. Date only (e.g. "2001-08-15").
-         */
-        expiration?: string
-
-        /**
-         * Optional. Whether the stop payment can stop more than one payment. False by default.
-         */
-        isMultiUse?: boolean
-
-        /**
-         * The description of the stop payment (up to 255 characters).
-         */
-        description: string
-
-        /**
-         * Optional. See [Tags](https://developers.unit.co/#tags).
-         */
-        tags?: Tags
-
-        /**
-         * Optional. See [Idempotency](https://docs.unit.co/#intro-idempotency).
-         */
-        idempotencyKey?: string
-    }
-    relationships: {
-        account: Relationship
-    }
-}
-
-export interface PatchAchStopPaymentRequest {
-    type: "achStopPayment"
-    attributes: {
-        /**
-         * Optional. See [Updating Tags](https://docs.unit.co/#updating-tags).
-         */
-        tags?: Tags
-
-        /**
-         * Optional. Date only (e.g. "2026-08-15").
-         */
-        expiration?: string
-    }
 }
 
 export interface ApproveCheckPaymentRequest {
