+++ types/index.ts
@@ -19,6 +19,7 @@ export * from "./institution"
 export * from "./orgToken"
 export * from "./payments"
 export * from "./checkPayment"
+export * from "./stopPayments"
 export * from "./repayments"
 export * from "./recurringPayment"
 export * from "./returns"
